#1 Allow users to specify the build source

+++ jdk11/build.sh
@@ -1,8 +1,22 @@
-cd "$(dirname "$0")"
+#!/bin/bash
+
+if [ "$1" = "" ]
+then
+    echo "no argument"
+    exit
+fi
+
+JDK_SOURCE=$1
+
+# clean up
 rm -rf build
 mkdir build
+
+# build
 docker pull tabatad/jdk11:latest
-docker run --name jdk11 tabatad/jdk11:latest
-docker cp jdk11:/jdk11u-dev/build/linux-x86_64-normal-server-release/images/jdk build/
+docker run -itd --name jdk11 --env JDK_SOURCE=$JDK_SOURCE tabatad/jdk11:latest
+docker cp ./src/$JDK_SOURE jdk11:/
+docker exec jdk11 /startup.sh
+docker cp jdk11:/src/$JDK_SOURCE/build/linux-x86_64-normal-server-release/images/jdk build/
 docker stop jdk11
 docker rm jdk11
